% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_postprocess.R
\name{aggregate_iter}
\alias{aggregate_iter}
\title{Compute posterior mean after burnin}
\usage{
aggregate_iter(x, nburn = -1, is_angle = FALSE, quantiles = c())
}
\arguments{
\item{x}{data}

\item{nburn}{number of burnin to skip, less than zero means half}

\item{is_angle}{are the values angles (uses vector averaging)}

\item{quantiles}{vector of quantiles to extact}
}
\description{
Compute posterior mean after burnin
}
