% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_model.R
\name{calc_ct_sum}
\alias{calc_ct_sum}
\title{Calculate sum of squared residuals from the state equations
corresponding to a single cycle according to equation (21) in
Harvey's paper (2005)}
\usage{
calc_ct_sum(order_trend, order_cycle, nth_freq, lambda, rho, n, alpha_draw)
}
\arguments{
\item{order_trend}{order of the trend part of the model}

\item{order_cycle}{order of cycle part of model, assumed same for all freqs}

\item{nth_freq}{index of frequency being queried}

\item{lambda}{frequency}

\item{rho}{dampening factor}

\item{alpha_draw}{state vector, for instance from posterior draw in Gibbs}
}
\value{
ct the calculated sum squared residuals
}
\description{
Calculate sum of squared residuals from the state equations
corresponding to a single cycle according to equation (21) in
Harvey's paper (2005)
}
