% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_uncond_cov.R
\name{calc_cycle_covariance_matrix}
\alias{calc_cycle_covariance_matrix}
\title{Build unconditional covariance matrix for cycle of one frequency}
\usage{
calc_cycle_covariance_matrix(order_cycle, rho, lambda)
}
\arguments{
\item{order_cycle}{order of cycle}

\item{rho}{dampening factor}

\item{lambda}{frequency of cycle}
}
\value{
covariance matrix
}
\description{
Build unconditional covariance matrix for cycle of one frequency
}
