% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_uncond_cov.R
\name{calc_cycle_variance_matrix_ij}
\alias{calc_cycle_variance_matrix_ij}
\title{Calculate ij components for the unconditional covariance matrix
The coefficients are produced according to equation (8) in Trimbur (2003)
on higher orders stochastic cycles. What is difference with prior?}
\usage{
calc_cycle_variance_matrix_ij(i, j, order_cycle, rho, lambda)
}
\arguments{
\item{i}{row of matrix}

\item{j}{column of matrix}

\item{order_cycle}{the order (n) of the stochastic cycle model}

\item{rho}{dampening factor}

\item{lambda}{frequency of the cycle}
}
\value{
ij entry of matrix
}
\description{
Calculate ij components for the unconditional covariance matrix
The coefficients are produced according to equation (8) in Trimbur (2003)
on higher orders stochastic cycles. What is difference with prior?
}
