% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_crossvalid.R
\name{calculate_blocks2}
\alias{calculate_blocks2}
\title{Calculate a list of blocks used for validation}
\usage{
calculate_blocks2(
  ylen,
  foldsize = 4 * 36,
  blocksize = 1008,
  first = 1,
  lastmax = -1
)
}
\arguments{
\item{ylen}{observed data length.}

\item{foldsize}{size of the folds (gaps) that are deleted}

\item{blocksize}{spacing of the blocks. Agap is created every blocksize}

\item{first}{first index to include}

\item{lastmax}{maximum last index to include}
}
\value{
a list of indices to drop for each step of the cross validation
}
\description{
Calculate a list of blocks used for validation
}
