% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_filter_response.R
\name{calculate_gain}
\alias{calculate_gain}
\title{Evaluate filter gain at requested frequencies}
\usage{
calculate_gain(evalfreqs, freqc, m, n, rho, qzeta, qkappa)
}
\arguments{
\item{evalfreqs}{vector of frequencies where gain is requested}

\item{freqc}{list of frequencies for cycles}

\item{m}{order of trend}

\item{n}{order of cycle}

\item{rho}{dampening}

\item{qzeta}{variance of trend divided by variance of observation}

\item{qkappa}{variance of cycle divided by variance of observation}
}
\value{
vector of gains
}
\description{
Evaluate filter gain at requested frequencies
}
