% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_postprocess.R
\name{colModes}
\alias{colModes}
\title{Take the mode of every column}
\usage{
colModes(mat1)
}
\arguments{
\item{mat1}{matrix over which modes are taken}
}
\value{
vector of modes
}
\description{
Take the mode of every column
}
