% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_postprocess.R
\name{do_plot}
\alias{do_plot}
\title{do plot}
\usage{
do_plot(ysave, y, recon, select, harmonic = 0)
}
\arguments{
\item{ysave}{saved data in case y has gaps}

\item{y}{data}

\item{recon}{reconstructed state}

\item{select}{selection to plot}
}
\description{
do plot
}
