% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_filter_response.R
\name{example_filter}
\alias{example_filter}
\title{Example of filter prior/loss function evaluation}
\usage{
example_filter(
  eps,
  zeta,
  kappa,
  rho,
  m = 2,
  n = 4,
  scale = 0.001,
  nsample = 2400
)
}
\value{
cost function evaluated
}
\description{
Example of filter prior/loss function evaluation
}
