% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_filter_response.R
\name{filter_prior}
\alias{filter_prior}
\title{Evaluate improper prior representing filter quality}
\usage{
filter_prior(
  test_freqs,
  test_targets,
  scale,
  sc_freq,
  m,
  n,
  rho,
  qzeta,
  qkappa
)
}
\arguments{
\item{test_freqs}{list of frequencies at which filter quality is to be eval'd}

\item{test_targets}{ideal gain at test freqs (1 or 0)}

\item{scale}{scale of penalty. The smaller, the less forgiving}

\item{sc_freq}{list of frequencies for cycles}

\item{m}{order of trend}

\item{n}{order of cycle}

\item{rho}{dampening}

\item{qzeta}{variance of trend divided by variance of observation}

\item{qkappa}{variance of cycle divided by variance of observation}
}
\value{
cost function evaluated
}
\description{
Evaluate improper prior representing filter quality
}
