% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_model.R
\name{find_cycle_index}
\alias{find_cycle_index}
\title{Return the last index of a stochastic cycle}
\usage{
find_cycle_index(order_trend, order_cycle, nth_freq)
}
\arguments{
\item{order_trend}{order of the trend part of the model}

\item{order_cycle}{order of cycle part of model, assumed same for all freqs}

\item{nth_freq}{index of frequency being queried}
}
\description{
Return the last index of a stochastic cycle
}
