% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_postprocess.R
\name{interpret_state}
\alias{interpret_state}
\title{Extract amplitudes, phase and values from state output matrix}
\usage{
interpret_state(state, order_trend, order_cycle, nfreq, flabels)
}
\arguments{
\item{state}{state to interpret}

\item{order_trend}{order of trend}

\item{order_cycle}{order of cycle}

\item{nfreq}{number of frequencies}

\item{flabels}{names for freqs}
}
\description{
Extract amplitudes, phase and values from state output matrix
}
