% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_examples.R
\name{jay3}
\alias{jay3}
\title{Jay(1999) tide with subtide interference}
\usage{
jay3(n = 4, freqs = c("M2", "O1", "K1", "SUB"), nhour = 805, samples_hr = 4)
}
\arguments{
\item{n}{cycles per day in subtide}

\item{freqs}{vector of frequencies to include (M2,O1,K1,SUB)}
}
\description{
Jay(1999) tide with subtide interference
}
