% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirp.R
\name{jay_flinchem_chirptest}
\alias{jay_flinchem_chirptest}
\title{Fit MCMC/Gibbs model
Mostly fits the Harvey and Trimbur model, except as noted here where some
parameters are more constrained}
\usage{
jay_flinchem_chirptest(scale_A = 1.1, scale_Aj1 = 1.24, scale_gamma = 1)
}
\arguments{
\item{scale_A}{Scaling of amplitude}

\item{scale_Aj1}{Scaling of amplitude response to Qr}

\item{scale_gamma}{Scaling of phase fluctuation}
}
\description{
Fit MCMC/Gibbs model
Mostly fits the Harvey and Trimbur model, except as noted here where some
parameters are more constrained
}
