% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_examples.R
\name{jay_subtide_interfere}
\alias{jay_subtide_interfere}
\title{Jay(1999) tide with subtide interference}
\usage{
jay_subtide_interfere(n = 4)
}
\arguments{
\item{n}{cycles per day in subtide}
}
\description{
Jay(1999) tide with subtide interference
}
