% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_filter_response.R
\name{plot_gain}
\alias{plot_gain}
\title{Plot gains of the sc components}
\usage{
plot_gain(sc_freq, m, n, rho, qzeta, qkappa, nsample)
}
\arguments{
\item{sc_freq}{list of frequencies for cycles}

\item{m}{order of trend}

\item{n}{order of cycle}

\item{rho}{dampening}

\item{qzeta}{variance of trend divided by variance of observation}

\item{qkappa}{variance of cycle divided by variance of observation}

\item{nsample}{number of evaluaiton points up to 5cpd #todo: flexible}
}
\value{
vector of gains
}
\description{
Plot gains of the sc components
}
