% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_postprocess.R
\name{reconstruct_state}
\alias{reconstruct_state}
\title{recreate full state}
\usage{
reconstruct_state(state, order_trend, order_cycle, nfreq)
}
\arguments{
\item{state}{s}

\item{order_trend}{order of trend}

\item{order_cycle}{order of cycle}

\item{nfreq}{number of frequencies}
}
\description{
recreate full state
}
