% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_gibbs.R
\name{sc_gibbs}
\alias{sc_gibbs}
\title{Fit MCMC/Gibbs model
Mostly fits the Harvey and Trimbur model, except as noted here where some
parameters are more constrained}
\usage{
sc_gibbs(
  y,
  order_trend,
  order_cycle,
  freqs,
  regress,
  regress_lmr,
  regress_freq,
  use_rho_prior,
  rho_high,
  rho_low,
  mh_sd_rho,
  sigma2_zeta_fixed,
  zeta_prior_rate,
  zeta_prior_shape,
  initial_sigma2_zeta,
  initial_sigma2_kappa,
  sigma2_diffuse,
  mc_iter,
  thin_rate
)
}
\arguments{
\item{y}{The observed data}

\item{order_trend}{order of the trend (low pass) component of the model}

\item{order_cycle}{order of the cyclical part of the model}

\item{freqs}{frequencies used in fitting}

\item{regress}{use regression?}

\item{regress_lmr}{use offline regression?}

\item{regress_freq}{regression frequencies}

\item{use_rho_prior}{whether to use a prior for rho (multiplied by uniform prior)}

\item{rho_high}{upper bound on rho, if same as rho_low to 5 digits rho will be fixed}

\item{rho_low}{lower bound on rho uniform. Initial rho is average of rho_high and rho_low}

\item{mh_sd_rho}{Metropolis Hastings random walk std. deviation for rho}

\item{sigma2_zeta_fixed}{fixed value or anything <0 if disabled}

\item{zeta_prior_rate}{a}

\item{zeta_prior_shape}{b}

\item{initial_sigma2_zeta}{c}

\item{initial_sigma2_kappa}{d}

\item{sigma2_diffuse}{magnitude for diffuse initial condition}

\item{mc_iter}{number of iterations}

\item{thin_rate}{thinning rate for posterior draws (every thin_rate sample is used)}
}
\description{
Fit MCMC/Gibbs model
Mostly fits the Harvey and Trimbur model, except as noted here where some
parameters are more constrained
}
