% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_gibbs_filter_prior.R
\name{sc_gibbs_filter_prior}
\alias{sc_gibbs_filter_prior}
\title{Fit MCMC/Gibbs model
Mostly fits the Harvey and Trimbur model, except as noted here where some
parameters are more constrained}
\usage{
sc_gibbs_filter_prior(
  y,
  order_trend,
  order_cycle,
  freqs,
  test_freqs,
  test_targets,
  filter_scale,
  samples_per_hr,
  regress,
  regress_lmr,
  regress_freq,
  initial_sigma2_epsilon,
  use_rho_prior,
  rho_initial,
  mh_sd_rho,
  max_rho,
  sigma2_zeta_fixed,
  zeta_prior_rate,
  zeta_prior_shape,
  mh_sd_zeta,
  mh_sd_eps,
  initial_sigma2_zeta,
  initial_sigma2_kappa,
  mh_sd_kappa,
  sigma2_diffuse,
  mc_iter,
  thin_rate
)
}
\arguments{
\item{y}{The observed data}

\item{order_trend}{order of the trend (low pass) component of the model}

\item{order_cycle}{order of the cyclical part of the model}

\item{freqs}{frequencies used in fitting}

\item{test_freqs}{filter quality test freqs in cpd}

\item{test_targets}{filter quality test targets (gain)}

\item{filter_scale}{scale of loss/prior on bad filter fit to freqs/targets}

\item{samples_per_hr}{number of obs per hour}

\item{regress}{use regression?}

\item{regress_lmr}{use offline regression?}

\item{regress_freq}{regression frequencies}

\item{use_rho_prior}{whether to use a prior for rho (multiplied by uniform prior)}

\item{rho_initial}{starting rho}

\item{mh_sd_rho}{Metropolis Hastings random walk std. deviation for rho}

\item{max_rho}{top of the uniform prior, often guided by conditioning of problem}

\item{sigma2_zeta_fixed}{fixed value or anything <0 if disabled}

\item{zeta_prior_rate}{a}

\item{zeta_prior_shape}{b}

\item{mh_sd_zeta}{Metropolis Hastings random walk std. deviation for zeta}

\item{initial_sigma2_zeta}{c}

\item{initial_sigma2_kappa}{initial variance for kappa, either scalar or one per freq}

\item{mh_sd_kappa}{Metropolis Hastings random walk std. deviation for kappa, either scalar or one per freq}

\item{sigma2_diffuse}{magnitude for diffuse initial condition}

\item{mc_iter}{number of iterations}

\item{thin_rate}{thinning rate for posterior draws (every thin_rate sample is used)}
}
\description{
Fit MCMC/Gibbs model
Mostly fits the Harvey and Trimbur model, except as noted here where some
parameters are more constrained
}
