% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_model.R
\name{sc_model_build}
\alias{sc_model_build}
\title{Calculate the filter lengths empirically using a unit impulse}
\usage{
sc_model_build(
  order_trend,
  order_cycle,
  freqs,
  rho,
  sigma2_eps,
  sigma2_zeta,
  sigma2_kappa
)
}
\arguments{
\item{order_trend}{Order of the trend (low pass) component}

\item{order_cycle}{Order of the cyclical component, same for all frequencies}

\item{freqs}{frequencies deployed}

\item{rho}{dampening factor}

\item{sigma2_eps}{variance of observation}

\item{sigma2_zeta}{variance of trend}

\item{sigma2_kappa}{variance of cycle disturbances}
}
\value{
list of approximate trend extraction filter lengths (2*sigma)
}
\description{
Calculate the filter lengths empirically using a unit impulse
}
