% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_postprocess.R
\name{sc_model_from_parms}
\alias{sc_model_from_parms}
\title{Constructs the matrices \code{FF} and \code{GG} used by DLM to represent
the stochastic cycle plus trend model}
\usage{
sc_model_from_parms(parms)
}
\arguments{
\item{parms}{list of parameters, possibly output of sc_params_from_gibbs}
}
\value{
model dlm model
}
\description{
Constructs the matrices \code{FF} and \code{GG} used by DLM to represent
the stochastic cycle plus trend model
}
