% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_model.R
\name{sc_model_mat}
\alias{sc_model_mat}
\title{Build the stochastic cycle model plus trend matrices}
\usage{
sc_model_mat(order_trend, order_cycle, freqs, rho, trend_scale = 1)
}
\arguments{
\item{order_trend}{Order of the trend (low pass) component}

\item{order_cycle}{Order of the cyclical component, same for all frequencies}

\item{freqs}{frequencies deployed}

\item{rho}{dampening factor}

\item{trend_scale}{scaling factor trend to observation}
}
\value{
list with attributes FF and GG
}
\description{
Constructs the matrices \code{FF} and \code{GG} used by DLM to represent
the stochastic cycle plus trend model
}
