% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_postprocess.R
\name{sc_params_from_gibbs}
\alias{sc_params_from_gibbs}
\title{Extract params from Gibbs sc_out output}
\usage{
sc_params_from_gibbs(gibbs_out, nburn = -1)
}
\arguments{
\item{nburn}{number of cycles to discard as MCMC burn-in. If <0, first half is discarded}

\item{gibbs_output}{output of sc_gibbs}
}
\value{
sc_params Posterior mean parameters from model
}
\description{
Constructs the matrices \code{FF} and \code{GG} used by DLM to represent
the stochastic cycle plus trend model
}
