% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_filter_response.R
\name{sparse_filter_targets}
\alias{sparse_filter_targets}
\title{Return a list of targets and frequencies that is good for D1-D2-D3-D4}
\usage{
sparse_filter_targets()
}
\value{
list of testing frequencies and ideal filter targets (0 or 1)
}
\description{
Return a list of targets and frequencies that is good for D1-D2-D3-D4
}
