% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_postprocess.R
\name{summarize_scout_result}
\alias{summarize_scout_result}
\title{Summarize the state info from Gibbs run and interpretation of run}
\usage{
summarize_scout_result(
  sc_out,
  interp,
  out_dt = 0.25,
  nburn = 10,
  outfile = NULL
)
}
\arguments{
\item{sc_out}{Result of Bayes computation}

\item{interp}{Single pass result from interpret_state}

\item{out_dt}{Time step to use for labeling output (e.g. 0.25 for 15min)}

\item{nburn}{number of burnin to skip, less than zero means half}

\item{outfile}{Name of output file for resulting dataframe (or NULL)}
}
\description{
Summarize the state info from Gibbs run and interpretation of run
}
